refactor : placeDetailHeader 분리 (#104)
